(PUP-4622) Add a diff command in Puppet filebucket
This patch adds a diff subcommand to the filebucket command.
It implements a way to diff filebucket files. Either between
them or between a filebucket file and the local one.

The diff command behaves like a real diff. If the second
argument is ommited, it will to try to find the original path
(and only one).

Modified the tests accordingly.

Conflicts:
	spec/unit/indirector/file_bucket_file/file_spec.rb