(PUP-4760) Service type should not retrieve ensure status when not needed
Before this, service type will always retrieve the ensure status even if no
ensure parameters is given.

But, because no Service Type parameter is dependant on whether the service
is running or not, we can just skip the ensure status retrieval.

Now, a service type will only retrieve the ensure status when a ensure
parameter is given.